Import yaml representers from own class
