+++ scripts/spack_manifest/injection/modules.py
@@ -10,19 +10,13 @@ from scripts.spack_manifest.getter import (
     Projections,
     Specs
 )
+from scripts.spack_manifest.injection.yaml_representer import (
+    YamlExplicitFlowStyleSequence,
+    yaml_explicit_flow_style_sequence_representer,
+    enforce_explicit_flow_style_definitions
+)
 
-# For sequence definitions, we keep it in the flow-style format (i.e., [a, b, c]) rather than block style
-# as it is more compact for reserved definitions.
-class YamlExplicitFlowStyleSequence(list[str]):
-    pass
-
-def yaml_explicit_flow_style_sequence_representer(dumper, data):
-    """
-    Custom representer for YAML to ensure that some sequences are represented in flow style.
-    This is necessary for sequences that are used as definitions in spack manifests.
-    """
-    return dumper.represent_sequence("tag:yaml.org,2002:seq", data, flow_style=True)
-
+# We represent reserved definitions as in flow-style sequences (eg. `[a]` rather than `- a`), so it is more compact.
 yaml.add_representer(YamlExplicitFlowStyleSequence, yaml_explicit_flow_style_sequence_representer)
 
 ##################
@@ -68,23 +62,6 @@ def main():
         with open(args.output, "w") as output_file:
             output_file.write(dumped_manifest)
 
-def enforce_explicit_flow_style_definitions(manifest: dict[str, Any]) -> dict[str, Any]:
-    """
-    Ensure that the 'definitions' section of the manifest is represented in flow style.
-    This is necessary for spack manifests to ensure that definitions are correctly interpreted.
-    """
-    if "spack" in manifest and "definitions" in manifest["spack"]:
-        definitions: list[dict[str, Any]] = manifest["spack"]["definitions"]
-        for i in range(len(definitions)):
-            definition = definitions[i]
-            if len(definition) > 0:
-                reserved_definition, reserved_value_list = list(definition.items())[0]
-
-                if reserved_definition.startswith("_"):
-                    manifest["spack"]["definitions"][i][reserved_definition] = YamlExplicitFlowStyleSequence(reserved_value_list)
-
-    return manifest
-
 def inject_projections(
     manifest: str, root_spec: str, packages: set[str]
 ) -> dict[str, Any]:
@@ -108,7 +85,6 @@ def inject_projections(
     # To start with, add the projections that are already defined in the manifest
     new_projections: dict[str, str] = dict(defined_projections_dict)
 
-    # if root_spec not in defined_projections:
     new_projections.update(
         generate_projection_for_root_spec_or_raise(manifest, root_spec, defined_projections_dict.get(root_spec))
     )
